third_party/blink/web_tests/fast/lists/ol-nested-items-expected.txt

This tests that list items are properly marked when some of the items are indirect children of the list element.

1 One
2 Two
3 Three
4 Four
5 Five
-> 6 Six
10 Ten
11 Eleven
-> 12 Twelve
-> 13 Thirteen
14 Fourteen
15 Fifteen
